requirements.txt
pyyaml
pyserial
pandas
flask